Rename logging.py to logger.py
Import rules from Python2 makes this a bad idea
