+++ tools/train_net.py
@@ -22,7 +22,7 @@ from maskrcnn_benchmark.utils.checkpoint import DetectronCheckpointer
 from maskrcnn_benchmark.utils.collect_env import collect_env_info
 from maskrcnn_benchmark.utils.comm import synchronize
 from maskrcnn_benchmark.utils.imports import import_file
-from maskrcnn_benchmark.utils.logging import setup_logger
+from maskrcnn_benchmark.utils.logger import setup_logger
 from maskrcnn_benchmark.utils.miscellaneous import mkdir
